Feature/ma 72

+++ backbone/src/main/java/org/researchstack/backbone/ui/ActiveTaskActivity.java
@@ -1,9 +1,17 @@
 package org.researchstack.backbone.ui;
 
+import android.annotation.TargetApi;
 import android.content.Context;
 import android.content.Intent;
+import android.content.pm.ActivityInfo;
+import android.os.Build;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.view.Surface;
+import android.view.WindowManager;
 
+import org.researchstack.backbone.PermissionRequestManager;
+import org.researchstack.backbone.R;
 import org.researchstack.backbone.result.FileResult;
 import org.researchstack.backbone.result.StepResult;
 import org.researchstack.backbone.result.logger.DataLoggerManager;
@@ -11,6 +19,10 @@ import org.researchstack.backbone.step.Step;
 import org.researchstack.backbone.step.active.ActiveStep;
 import org.researchstack.backbone.step.active.CountdownStep;
 import org.researchstack.backbone.task.Task;
+import org.researchstack.backbone.ui.callbacks.ActivityCallback;
+import org.researchstack.backbone.ui.step.layout.ActiveStepLayout;
+import org.researchstack.backbone.ui.step.layout.StepLayout;
+import org.researchstack.backbone.ui.step.layout.StepPermissionRequest;
 
 import java.io.File;
 import java.util.ArrayList;
@@ -26,7 +38,7 @@ import java.util.Map;
  * and make sure that they are correctly bundled and uploaded at the end
  */
 
-public class ActiveTaskActivity extends ViewTaskActivity {
+public class ActiveTaskActivity extends ViewTaskActivity implements ActivityCallback {
 
     private boolean isBackButtonEnabled;
 
@@ -39,19 +51,21 @@ public class ActiveTaskActivity extends ViewTaskActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
-        if(savedInstanceState == null) {
-            init();  // for the first time
-        }
+        init();
     }
 
     protected void init() {
-        DataLoggerManager.initialize(this);
-        DataLoggerManager.getInstance().deleteAllDirtyFiles();
+        if (!DataLoggerManager.isInitialized()) {
+            DataLoggerManager.initialize(this);
+            DataLoggerManager.getInstance().deleteAllDirtyFiles();
+        }
     }
 
     @Override
     protected void discardResultsAndFinish() {
+        if (currentStepLayout instanceof ActiveStepLayout) {
+            ((ActiveStepLayout) currentStepLayout).forceStop();
+        }
         DataLoggerManager.getInstance().deleteAllDirtyFiles();
         super.discardResultsAndFinish();
     }
@@ -62,7 +76,7 @@ public class ActiveTaskActivity extends ViewTaskActivity {
         // compute back button status while currentStep is actually the previousStep at this point
         isBackButtonEnabled =
                 (!(step instanceof ActiveStep) || (step instanceof CountdownStep)) &&
-                        !(currentStep instanceof ActiveStep); // currentStep is one previously showing at this point
+                 !(currentStep instanceof ActiveStep); // currentStep is one previously showing at this point
         if (getSupportActionBar() != null) {
             getSupportActionBar().setDisplayHomeAsUpEnabled(isBackButtonEnabled);
         }
@@ -71,6 +85,17 @@ public class ActiveTaskActivity extends ViewTaskActivity {
         // unnecessarily, so if it already exists and is showing, then do not re-show the StepLayout
         if (!isStepAnAlreadyShowingActiveStep(step)) {
             super.showStep(step, alwaysReplaceView);
+        }
+
+        // Active steps lock screen on and orientation so that the view is not unnecessarily
+        // destroyed and recreated while the data logger is recording
+        // TODO: do we need a partial CPU wake lock here?
+        if (step instanceof ActiveStep) {
+            lockScreenOn();
+            lockOrientation();
+        } else {
+            unlockScreenOn();
+            unlockOrientation();
         }
     }
 
@@ -87,7 +112,17 @@ public class ActiveTaskActivity extends ViewTaskActivity {
     }
 
     @Override
+    public void setRequestedOrientation(int requestedOrientation) {
+        super.setRequestedOrientation(requestedOrientation);
+    }
+
+    @Override
     protected void saveAndFinish() {
+
+        // just in case we were locking these
+        unlockScreenOn();
+        unlockOrientation();
+
         taskResult.setEndDate(new Date());
 
         // Loop through and find all the FileResult files
@@ -117,5 +152,85 @@ public class ActiveTaskActivity extends ViewTaskActivity {
 
         // TODO: move this to the successful/failure block of the upload web service call
         super.saveAndFinish();
+    }
+
+    /**
+     * Active Steps lock screen to on so it can avoid any interruptions during data logging
+     */
+    private void lockScreenOn() {
+        getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+    }
+
+    private void unlockScreenOn() {
+        getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+    }
+
+    /**
+     * Active Steps lock orientation so it can avoid any interruptions during data logging
+     */
+    private void lockOrientation() {
+        int orientation;
+        int rotation = ((WindowManager) getSystemService(
+                Context.WINDOW_SERVICE)).getDefaultDisplay().getRotation();
+        switch (rotation) {
+            case Surface.ROTATION_0:
+                orientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT;
+                break;
+            case Surface.ROTATION_90:
+                orientation = ActivityInfo.SCREEN_ORIENTATION_LANDSCAPE;
+                break;
+            case Surface.ROTATION_180:
+                orientation = ActivityInfo.SCREEN_ORIENTATION_REVERSE_PORTRAIT;
+                break;
+            default:
+                orientation = ActivityInfo.SCREEN_ORIENTATION_REVERSE_LANDSCAPE;
+                break;
+        }
+        setRequestedOrientation(orientation);
+    }
+
+    private void unlockOrientation() {
+        setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_UNSPECIFIED);
+    }
+
+    @Override
+    protected void onExecuteStepAction(int action) {
+        // In this case, we cannot complete the Active Task, since one of the ActiveSteps
+        // Requested we end the task, because it couldn't complete for some reason
+        if (action == ACTION_END && currentStep instanceof ActiveStep) {
+            discardResultsAndFinish();
+        } else {
+            super.onExecuteStepAction(action);
+        }
+    }
+
+    @TargetApi(Build.VERSION_CODES.M)
+    @Override
+    public void onRequestPermission(String id) {
+        if (PermissionRequestManager.getInstance().isNonSystemPermission(id)) {
+            PermissionRequestManager.getInstance().onRequestNonSystemPermission(this, id);
+        } else {
+            requestPermissions(new String[] {id}, PermissionRequestManager.PERMISSION_REQUEST_CODE);
+        }
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        if(requestCode == PermissionRequestManager.PERMISSION_REQUEST_CODE) {
+            updateStepLayoutForPermission();
+        }
+    }
+
+    protected void updateStepLayoutForPermission() {
+        StepLayout stepLayout = (StepLayout) findViewById(R.id.rsb_current_step);
+        if(stepLayout instanceof StepPermissionRequest) {
+            ((StepPermissionRequest) stepLayout).onUpdateForPermissionResult();
+        }
+    }
+
+    @Override
+    public void startConsentTask() {
+        // deprecated
     }
 }
